requirements/base.txt
Django==5.1.3
environs==11.2.0
psycopg2==2.9.10
environs==11.2.1
django-ckeditor==6.7.2
django-mptt==0.16.0
djangorestframework==3.15.2

drf-spectacular==0.28.0
